Generate a grammar.ts
This reverts cc6bb1cd and instead adds a script to bend the generated grammar.js file to a somewhat valid grammar.ts file.

We do that because we need to publish both ESM and CJS, and transpiling from a .ts is by far the most reliable way to do so.