Introduce cel-spec package and update to cel-spec v0.18.0
The cel-spec defines a Protobuf schema for the CEL AST, and various types for evaluation. It also defines a conformance test suites, as a set of proto files and test data.

This adds the package `@bufbuild/cel-spec` to the repository. The purpose of this package is to provide generated files for all proto files, as well as conformance test data.

The proto files and test data in this package are fetched from the upstream repository github.com/google/cel-spec. A git reference (tag, commit, or branch) is specified in package.json, field "upstreamCelSpecRef".