Prepare for protobuf-es v2
Avoid Message.getType because it will be unavailable. Update some switch statements switching on message type to switch on type name instead.

The is one usage of getType left in CelAdapter. It's used to implement CelValAdapter.equals, and still needs to be replaced.

This also reduces visibility for some classes and methods to help refactoring.