Bump target (and library) from ES2020 to ES2023 (#230)
Our expectation is that bundlers will transpile syntax features to compatible expressions. But bundlers will typically not automatically provide API polyfills. This is a concern for users targeting an older standard - they have to provide their own polyfills.

This is unfortunate, but the alternative is that we spend considerable time working around API limitations instead of improving CEL. Given that this implementation is new, our hope is that ES2023 is very widely available by the time it's mature. Feedback from CEL users is welcome.